meta(clippy): Replace from_str() with new()
clippy warns that from_str may be confused with FromStr being
implemented, so for simplicity new() is used instead. No alternative
instantiation methods are planned, which means that &str must be used if
users want to create Johnny.Decimal structs.

Note that new() is preferred over try_new() for simplicity and ease of
typing.