docs: Clarify project goals in README
Using a Rust web framework like Rocket introduces a lot of friction in
the development process. The current plan is to use React and possibly
Tauri to create a desktop application for non-CLI users.

Ultimately one of the advantages of the Johnny Decimal system is that
you can use pretty much any tool you want with it, as it's just a way to
organize data.